requirements.txt
av==12.3.0
black==24.8.0
cityscapesScripts==2.2.2
cloudpickle==3.0.0
docutils==0.20.1
fairscale==0.4.13
fvcore==0.1.5.post20221221
hydra-core==1.3.2
imagecorruptions==1.1.2
iopath==0.1.9
matplotlib==3.7.3
mmcv==2.2.0
mmdet==3.3.0
numpy==1.24.3
omegaconf==2.3.0
onnx==1.16.2
opencv_python==4.8.1.78
pandas==2.0.3
Pillow==9.3.0
psutil==5.9.8
pycocotools==2.0.7
Pygments==2.18.0
pytorch_lightning
PyYAML==6.0.1
recommonmark==0.7.1
Requests==2.32.3
scipy==1.10.0
setuptools==58.0.4
Shapely==2.0.5
Sphinx==7.4.0
sphinx_rtd_theme==2.0.0
tabulate==0.9.0
termcolor==2.4.0
timm==1.0.8
torch==1.9.0+cu111
torchvision==0.10.0+cu111
tqdm==4.66.1
scikit-image==0.18.3